+++ lib/renderer/json.rb
@@ -7,7 +7,7 @@ module Renderer
       @config = config
     end
 
-    def add_node(name)
+    def add_node(name, opts)
       @nodes << {id: name, label: name}
     end
 
@@ -17,7 +17,7 @@ module Renderer
 
     def output
       IO.write(@file_name, {nodes: @nodes, edges: @edges}.to_json)
+      Renderer.copy_asset('view.html', @file_name)
     end
-
   end
 end