+++ spec/integration/visualize_aws_spec.rb
@@ -29,7 +29,18 @@ describe VisualizeAws do
 
     it 'should parse json input', :integration => true do
       expect(FileUtils).to receive(:copy)
-      VisualizeAws.new(config, opts).unleash(temp_file.path)
+      VisualizeAws.new(config, opts.merge(:renderer => 'json')).unleash(temp_file.path)
+      expect(JSON.parse(expected_content)).to eq(JSON.parse(actual_content))
+    end
+  end
+
+  context 'json to navigator file' do
+    let(:expected_file) { File.join(File.dirname(__FILE__), 'navigator.json') }
+    let(:temp_file) { Tempfile.new(%w(aws .json)) }
+
+    it 'should parse json input', :integration => true do
+      expect(FileUtils).to receive(:copy)
+      VisualizeAws.new(config, opts.merge(:renderer => 'navigator')).unleash(temp_file.path)
       expect(JSON.parse(expected_content)).to eq(JSON.parse(actual_content))
     end
   end